requirements.txt
Flask
scikit-learn
scipy
pandas
catboost
xgboost
seaborn
matplotlib
dill
PyYAML
-e .